+++ packages/app-config-writer/src/preview-config/index.ts
@@ -1,8 +1,8 @@
 import { checkPrerequisites, getExplicitApprovalToAdjustFiles } from './prerequisites';
 import { create as createStorage } from 'mem-fs';
 import { create, type Editor } from 'mem-fs-editor';
-import { deleteNoLongerUsedFiles, renameDefaultSandboxes } from './preview-files';
-import { updatePreviewMiddlewareConfigs } from './ui5-yaml';
+import { deleteNoLongerUsedFiles, renameDefaultSandboxes, renameDefaultTestFiles } from './preview-files';
+import { updatePreviewMiddlewareConfigs, updateDefaultTestConfig } from './ui5-yaml';
 import { updateVariantsCreationScript } from './package-json';
 import { type ToolsLogger } from '@sap-ux/logger';
 
@@ -14,14 +14,19 @@ import { type ToolsLogger } from '@sap-ux/logger';
  * Corresponding files which are used for the preview are renamed or deleted.
  *
  * @param basePath - base path to be used for the conversion
- * @param logger logger to report info to the user
- * @param fs - file system reference
+ * @param options - options for the conversion
+ * @param options.convertTests - if set to true, then test suite and test runners fill be included in the conversion
+ * @param options.logger - logger to report info to the user
+ * @param options.fs - file system reference
  * @returns file system reference
  */
-export async function convertToVirtualPreview(basePath: string, logger?: ToolsLogger, fs?: Editor): Promise<Editor> {
-    if (!fs) {
-        fs = create(createStorage());
-    }
+export async function convertToVirtualPreview(
+    basePath: string,
+    options: { convertTests?: boolean; logger?: ToolsLogger; fs?: Editor }
+): Promise<Editor> {
+    const fs = options.fs ?? create(createStorage());
+    const logger = options.logger;
+    const convertTests = options.convertTests ?? false;
 
     if (!(await checkPrerequisites(basePath, fs, logger))) {
         throw Error('The prerequisites are not met. For more information, see the log messages above.');
@@ -32,9 +37,13 @@ export async function convertToVirtualPreview(basePath: string, logger?: ToolsLo
         return fs;
     }
 
-    await updatePreviewMiddlewareConfigs(fs, basePath, logger);
-    await renameDefaultSandboxes(fs, basePath);
-    await deleteNoLongerUsedFiles(fs, basePath);
+    await updatePreviewMiddlewareConfigs(fs, basePath, convertTests, logger);
+    await renameDefaultSandboxes(fs, basePath, logger);
+    if (convertTests) {
+        await renameDefaultTestFiles(fs, basePath, logger);
+        await updateDefaultTestConfig(fs, basePath, logger);
+    }
+    await deleteNoLongerUsedFiles(fs, basePath, logger);
     await updateVariantsCreationScript(fs, basePath, logger);
 
     return fs;