+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.4.42
+
+### Patch Changes
+
+-   Updated dependencies [e1edcd7]
+    -   @sap-ux/project-access@1.28.10
+
+## 0.4.41
+
+### Patch Changes
+
+-   b7c346c: fix: Moving Init.js functionality inline in flpSandbox.html
+
+## 0.4.40
+
+### Patch Changes
+
+-   1b3c47d: feat: Ensure UI5 2.x compliance
+
 ## 0.4.39
 
 ### Patch Changes