+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @sap-ux/inquirer-common
 
+## 0.6.1
+
+### Patch Changes
+
+-   Updated dependencies [fe0878d]
+    -   @sap-ux/fiori-generator-shared@0.7.18
+
+## 0.6.0
+
+### Minor Changes
+
+-   112d29a: Include the name label for default UI5 choice
+
+## 0.5.15
+
+### Patch Changes
+
+-   f8dda3b: Fix for multiple issues relating to prompt validations
+
+## 0.5.14
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.17
+-   @sap-ux/telemetry@0.5.50
+
+## 0.5.13
+
+### Patch Changes
+
+-   Updated dependencies [cb54b44]
+    -   @sap-ux/btp-utils@0.17.2
+    -   @sap-ux/fiori-generator-shared@0.7.16
+    -   @sap-ux/telemetry@0.5.49
+
+## 0.5.12
+
+### Patch Changes
+
+-   @sap-ux/telemetry@0.5.48
+-   @sap-ux/fiori-generator-shared@0.7.15
+
+## 0.5.11
+
+### Patch Changes
+
+-   0c64478: feat: Enhance extracted FLP prompts to support ADP scenario
+
 ## 0.5.10
 
 ### Patch Changes