+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @sap-ux/ui-prompting
 
+## 0.3.11
+
+### Patch Changes
+
+-   Updated dependencies [c570683]
+    -   @sap-ux/ui-components@1.23.0
+
+## 0.3.10
+
+### Patch Changes
+
+-   Updated dependencies [81a5351]
+    -   @sap-ux/ui-components@1.22.3
+
+## 0.3.9
+
+### Patch Changes
+
+-   Updated dependencies [318477a]
+    -   @sap-ux/ui-components@1.22.2
+
+## 0.3.8
+
+### Patch Changes
+
+-   Updated dependencies [0fd924b]
+    -   @sap-ux/ui-components@1.22.1
+
+## 0.3.7
+
+### Patch Changes
+
+-   Updated dependencies [0f9d186]
+    -   @sap-ux/ui-components@1.22.0
+
+## 0.3.6
+
+### Patch Changes
+
+-   Updated dependencies [d167a73]
+    -   @sap-ux/ui-components@1.21.4
+
 ## 0.3.5
 
 ### Patch Changes