+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.3.30
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.1
+
+## 0.3.29
+
+### Patch Changes
+
+-   Updated dependencies [112d29a]
+    -   @sap-ux/inquirer-common@0.6.0
+
+## 0.3.28
+
+### Patch Changes
+
+-   Updated dependencies [f8dda3b]
+    -   @sap-ux/inquirer-common@0.5.15
+
+## 0.3.27
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.14
+
+## 0.3.26
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.13
+
+## 0.3.25
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.12
+
+## 0.3.24
+
+### Patch Changes
+
+-   Updated dependencies [0c64478]
+    -   @sap-ux/inquirer-common@0.5.11
+
 ## 0.3.23
 
 ### Patch Changes